Restructure Android
