+++ src/pages/integration-builder/builder/steps/frameworks/android/replacementCode/get-constructor-code-android.ts
