+++ src/pages/integration-builder/builder/steps/frameworks/android/stepContent/index.ts
@@ -1,0 +1,32 @@
+import { toSteps } from "../../../../utils";
+import * as configureDeepLink from "./configure-deeplink.mdx";
+import * as CustomAuthentication from "./custom-authentication.mdx";
+import * as installationAppManifest from "./install-app-manifest.mdx";
+import * as installationBuildGradle from "./install-build-gradle.mdx";
+import * as installationAndroid from "./installation.mdx";
+import * as instantiate from "./instantiateSDK.mdx";
+import * as multiFactorAuthentication from "./mfa.mdx";
+import * as registerApp from "./register-app.mdx";
+import * as setResultURL from "./setResultURL.mdx";
+import * as singleTop from "./singleTop.mdx";
+import * as triggeringLogin from "./triggering-login.mdx";
+import * as triggeringLogout from "./triggering-logout.mdx";
+import * as whiteLabeling from "./whitelabeling.mdx";
+
+const STEPS = toSteps({
+  installationAndroid,
+  installationBuildGradle,
+  installationAppManifest,
+  registerApp,
+  configureDeepLink,
+  singleTop,
+  instantiate,
+  setResultURL,
+  triggeringLogin,
+  triggeringLogout,
+  whiteLabeling,
+  CustomAuthentication,
+  multiFactorAuthentication,
+});
+
+export default STEPS;
