+++ src/pages/integration-builder/builder/steps/frameworks/android/stepContent/install-app-manifest.mdx
