+++ src/pages/integration-builder/builder/steps/frameworks/android/stepContent/instantiateSDK.mdx
