+++ src/pages/integration-builder/builder/steps/frameworks/android/stepContent/register-app.mdx
